requirements.txt
streamlit==1.29.0
pillow
numpy==1.26.2
opencv-python
tensorflow==2.14.0
scikit-image==0.22.0
exifread==3.0.0
deepface==0.0.79
torch==2.1.1
torchvision==0.16.1
huggingface-hub==0.19.4
transformers==4.35.2
matplotlib==3.8.2
scipy==1.11.4
requests==2.31.0
# C2PA and content credentials verification
c2pa==0.3.2
pyjumbf==0.2.1
cryptography==41.0.3
pycontent-authenticator==0.1.4
pytesseract==0.3.10
